internal: refactor macro error
