+++ crates/hir_def/src/body/tests.rs
@@ -89,67 +89,6 @@ mod m {
 }
 
 #[test]
-fn macro_diag_builtin() {
-    check_diagnostics(
-        r#"
-#[rustc_builtin_macro]
-macro_rules! env {}
-
-#[rustc_builtin_macro]
-macro_rules! include {}
-
-#[rustc_builtin_macro]
-macro_rules! compile_error {}
-
-#[rustc_builtin_macro]
-macro_rules! format_args {
-    () => {}
-}
-
-fn f() {
-    // Test a handful of built-in (eager) macros:
-
-    include!(invalid);
-  //^^^^^^^^^^^^^^^^^ could not convert tokens
-    include!("does not exist");
-  //^^^^^^^^^^^^^^^^^^^^^^^^^^ failed to load file `does not exist`
-
-    env!(invalid);
-  //^^^^^^^^^^^^^ could not convert tokens
-
-    env!("OUT_DIR");
-  //^^^^^^^^^^^^^^^ `OUT_DIR` not set, enable "run build scripts" to fix
-
-    compile_error!("compile_error works");
-  //^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ compile_error works
-
-    // Lazy:
-
-    format_args!();
-  //^^^^^^^^^^^^^^ no rule matches input tokens
-}
-        "#,
-    );
-}
-
-#[test]
-fn macro_rules_diag() {
-    check_diagnostics(
-        r#"
-macro_rules! m {
-    () => {};
-}
-fn f() {
-    m!();
-
-    m!(hi);
-  //^^^^^^ leftover tokens
-}
-      "#,
-    );
-}
-
-#[test]
 fn unresolved_macro_diag() {
     check_diagnostics(
         r#"
@@ -161,30 +100,3 @@ fn f() {
     );
 }
 
-#[test]
-fn dollar_crate_in_builtin_macro() {
-    check_diagnostics(
-        r#"
-#[macro_export]
-#[rustc_builtin_macro]
-macro_rules! format_args {}
-
-#[macro_export]
-macro_rules! arg {
-    () => {}
-}
-
-#[macro_export]
-macro_rules! outer {
-    () => {
-        $crate::format_args!( "", $crate::arg!(1) )
-    };
-}
-
-fn f() {
-    outer!();
-  //^^^^^^^^ leftover tokens
-}
-        "#,
-    )
-}
